fix: move STFT off radar thread — stops frozen display and maze hang on Surface Pro
Root cause: RadarBridge._on_frame() was running the full spectrogram STFT
computation (30-frame batch) on the radar background thread every 100 ms.
As the frame buffer filled, computation time exceeded the 100 ms frame
interval, starving the radar thread and causing:
  • Visualize tab: display froze after ~3 s, showing only static blobs
  • Maze / RoboSoccer: inference received stale frames, gestures ignored

Fix:
  • SpectrogramProcessor.push_frame_raw() — lightweight accumulator for
    the radar thread (no computation, just deque.append)
  • SpectrogramProcessor.get_streaming_result(n_cols) — computes STFT on
    demand; safe to call from any thread
  • RadarBridge._on_frame() now only calls push_frame_raw() + emits raw signal
  • RadarBridge adds a QTimer (_display_timer, 5 Hz / 200 ms) that calls
    get_streaming_result() on the main thread and emits frame_ready
  • SpectrogramWidget.update_frame() — gaussian_filter stays float32
    (no float64 round-trip → 2× faster, lower peak memory)

Net effect: radar thread is fully decoupled from STFT cost. Display updates
reliably at 5 Hz (same 20 cols/s scroll rate). TestTab / maze inference
always gets fresh frames regardless of display load.

+++ ui/spectrogram_widget.py
@@ -86,5 +86,7 @@ class SpectrogramWidget(pg.GraphicsLayoutWidget):
             self._buffer[:, self._col] = col
             self._col = (self._col + 1) % BUFFER_WIDTH
         display = np.roll(self._buffer, -self._col, axis=1)
-        display = gaussian_filter(display.astype(np.float64), sigma=[1.5, 0.8]).astype(np.float32)
+        # Keep float32 throughout — scipy gaussian_filter handles it natively
+        # and avoids the 2× memory + compute cost of a float64 round-trip.
+        display = gaussian_filter(display, sigma=[1.5, 0.8])
         self._img.setImage(display.T, autoLevels=False)
